backend url configuration when npm run build used
