+++ client/.env.production
@@ -1,0 +1,1 @@
+VUE_APP_MODE=production
