+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ParraiN7",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
