+++ client/src/feathers-client.ts
@@ -1,11 +1,10 @@
 
-import BACKEND_URL from '@/config';
 import feathers from '@feathersjs/feathers';
 import auth from '@feathersjs/authentication-client';
 import socketio from '@feathersjs/socketio-client';
-import io from 'socket.io-client';
 
-const socket = io(BACKEND_URL);
+import socket from '@/socket-config';
+
 const app = feathers();
 app.configure(socketio(socket));
 app.configure(auth({
